+++ .storybook/main.ts
@@ -32,7 +32,8 @@ const config: StorybookConfig = {
       button[id="removed-reveal"] {display: none; }
       .sidebar-item:has(button[id="buttons-button-rive"]) { display: none; }
       .sidebar-item:has(button[id="buttons-button-split"]) { display: none; }
-      .sidebar-item:has(button[id="layout-card"]) { display: none; }
+      div[id="wip"] > button:first-child::after { content: " 🚧 "; }
+      button[id^="wip-"]::after { content: " 🚧 "; }
     </style>
   `,